backend-python/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.34.0
pydantic==2.11.0
pydantic-settings==2.8.0
numpy==2.2.0
pandas==2.2.3
pyyaml==6.0.2
faiss-cpu==1.11.0
httpx==0.28.1
feedparser==6.0.11
tenacity==9.0.0
orjson==3.11.0
structlog==25.2.0
openai==1.68.2
sentence-transformers==3.4.1
pytest==8.3.4
pytest-asyncio==0.25.0
